+++ web/app/connectors/_merlins-connector/account/commands.js
@@ -6,7 +6,6 @@ import {makeRequest, makeFormEncodedRequest} from 'progressive-web-sdk/dist/util
 import {jqueryResponse} from 'progressive-web-sdk/dist/jquery-response'
 import {SubmissionError} from 'redux-form'
 
-import {fetchCustomerAddresses} from '../checkout/commands'
 import {getCookieValue, splitFullName} from '../../../utils/utils'
 import {getFormKey} from '../selectors'
 import {fetchPageData} from '../app/commands'
@@ -22,8 +21,14 @@ import {
     receiveWishlistUIData
 } from 'progressive-web-sdk/dist/integration-manager/account/results'
 import {receiveWishlistProductData} from 'progressive-web-sdk/dist/integration-manager/products/results'
+
 import {receiveCheckoutLocations} from 'progressive-web-sdk/dist/integration-manager/checkout/results'
-import {buildFormData, createAddressRequestObject} from './utils'
+import {
+    buildFormData,
+    createAddressRequestObject,
+    fetchCustomerAddresses
+} from './utils'
+
 import {jqueryAjaxWrapper, parseAddress} from '../utils'
 import {LOGIN_POST_URL, CREATE_ACCOUNT_POST_URL, getDeleteAddressURL} from '../config'
 import {setLoggedIn} from 'progressive-web-sdk/dist/integration-manager/results'
@@ -69,11 +74,7 @@ export const initAccountAddressPage = (url) => (dispatch) => { // eslint-disable
     return fetchCustomerAddresses()
         .then(({customer: {addresses}}) => {
             const parsedAddresses = addresses.map((address) => parseAddress(address))
-            const addressInfo = {
-                addresses: parsedAddresses.filter((address) => !address.default),
-                defaultAddress: parsedAddresses.filter((address) => address.default)[0]
-            }
-            return dispatch(receiveAccountAddressData(addressInfo))
+            return dispatch(receiveAccountAddressData(parsedAddresses))
         })
 }